Renamed 'entity' package to 'entities' for compliance with common practice
