+++ back-end/src/main/java/com/exadel/team3/backend/dao/UserRepository.java
@@ -1,6 +1,6 @@
 package com.exadel.team3.backend.dao;
 
-import com.exadel.team3.backend.entity.User;
+import com.exadel.team3.backend.entities.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
