+++ back-end/src/main/java/com/exadel/team3/backend/services/UserService.java
@@ -1,7 +1,7 @@
 package com.exadel.team3.backend.services;
 
-import com.exadel.team3.backend.entity.User;
-import com.exadel.team3.backend.entity.UserRole;
+import com.exadel.team3.backend.entities.User;
+import com.exadel.team3.backend.entities.UserRole;
 
 import java.util.List;
